+++ frontend/techpick/src/components/DragOverlay/FolderItemOverlay.tsx
