+++ frontend/techpick/src/components/DragOverlay/PickDragOverlayShadowList.tsx
