+++ frontend/techpick/src/components/DragOverlay/dragOverlay.css.ts
