+++ frontend/techpick/src/components/DragOverlay/folderItemOverlay.css.ts
