+++ frontend/techpick/src/components/DragOverlay/pickDragOverlayShadowList.css.ts
