+++ frontend/techpick/src/components/DragOverlay/pickRecordOverlay.css.ts
